pyproject.toml
[project]
name = "Football_Play_Project"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "ultralytics>=8.3.0",
  "opencv-python",
  "numpy",
  "pandas",
  "torch",
  "torchvision",
  "matplotlib",
  "scikit-learn",
  "pyyaml"
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
